fix: fix app not closing when controls window is closed
The display window has decorations removed so it can appear in a stream without clutter. This removes the "close", "minimize", and "fullscreen" buttons. The app needs to quit when the controls panel is closed so that the user is not left with a lone display panel without a close button in the top corner.